Move build_selection to vortex-serde
